moved CheckType to common domain
