+++ src/aws_lambda_python/common_domain/enum/check_type.py
