+++ tests/unit/test_mpic_response_builder.py
@@ -4,9 +4,8 @@ from aws_lambda_python.common_domain.check_response import CaaCheckResponse, Dcv
     DcvCheckResponseDetails
 
 from aws_lambda_python.mpic_coordinator.config.service_config import API_VERSION
-from aws_lambda_python.mpic_coordinator.domain.enum.check_type import CheckType
-from aws_lambda_python.mpic_coordinator.domain.mpic_response import BaseMpicResponse, MpicCaaResponse, \
-    AnnotatedMpicResponse
+from aws_lambda_python.common_domain.enum.check_type import CheckType
+from aws_lambda_python.mpic_coordinator.domain.mpic_response import BaseMpicResponse, AnnotatedMpicResponse
 from aws_lambda_python.mpic_coordinator.mpic_response_builder import MpicResponseBuilder
 from pydantic import TypeAdapter
